requirements.txt
Django==6.0.3
djangorestframework==3.15.2
psycopg2-binary==2.9.9
django-cors-headers==4.3.1